fix: JPA, H2 걷어냄, application-local.yaml 임시 복구, 논의 필요
