+++ src/main/kotlin/thatline/localup/common/entity/BaseMongoDbEntity.kt
@@ -1,0 +1,13 @@
+package thatline.localup.common.entity
+
+import org.springframework.data.annotation.Id
+import java.time.LocalDateTime
+
+abstract class BaseMongoDbEntity(
+    @Id
+    val id: String,
+
+    val createdDate: LocalDateTime,
+
+    val lastModifiedDate: LocalDateTime,
+)
