--- src/main/kotlin/thatline/localup/common/entity/mongodb/BaseMongoDbEntity.kt
@@ -1,15 +1,0 @@
-package thatline.localup.common.entity.mongodb
-
-import jakarta.persistence.Id
-import jakarta.persistence.MappedSuperclass
-import java.time.LocalDateTime
-
-@MappedSuperclass
-class BaseMongoDbEntity(
-    @Id
-    val id: String,
-
-    val createdDate: LocalDateTime,
-
-    val lastModifiedDate: LocalDateTime,
-)
