+++ src/main/kotlin/thatline/localup/common/property/EtcApiProperty.kt
@@ -20,12 +20,12 @@ class EtcApiProperty(
         ) {
             data class GetUltraSrtNcst(
                 val secondPath: String,
-                val dateType: String,
+                val dataType: String,
             )
 
             data class GetFcstVersion(
                 val secondPath: String,
-                val dateType: String,
+                val dataType: String,
             )
         }
     }
