+++ src/main/kotlin/thatline/localup/etcapi/restclient/EtcApiRestClient.kt
@@ -48,7 +48,7 @@ class EtcApiRestClient(
             .queryParam("serviceKey", etcApiProperty.kma.vilageFcstInfoService.serviceKey)
             .queryParam("pageNo", pageNo)
             .queryParam("numOfRows", numOfRows)
-            .queryParam("dataType", etcApiProperty.kma.vilageFcstInfoService.getUltraSrtNcst.dateType)
+            .queryParam("dataType", etcApiProperty.kma.vilageFcstInfoService.getUltraSrtNcst.dataType)
             .queryParam("base_date", baseDate)
             .queryParam("base_time", baseTime)
             .queryParam("nx", nx)
@@ -95,7 +95,7 @@ class EtcApiRestClient(
             .queryParam("serviceKey", etcApiProperty.kma.vilageFcstInfoService.serviceKey)
             .queryParam("pageNo", pageNo)
             .queryParam("numOfRows", numOfRows)
-            .queryParam("dataType", etcApiProperty.kma.vilageFcstInfoService.getFcstVersion.dateType)
+            .queryParam("dataType", etcApiProperty.kma.vilageFcstInfoService.getFcstVersion.dataType)
             .queryParam("ftype", fileType)
             .queryParam("basedatetime", basedDatetime)
             .build(true)
