--- src/main/kotlin/thatline/localup/storage/repository/mongodb/FileMongoDbRepository.kt
@@ -1,12 +1,0 @@
-package thatline.localup.storage.repository.mongodb
-
-import org.springframework.data.mongodb.repository.MongoRepository
-import org.springframework.stereotype.Repository
-import thatline.localup.storage.entity.mongodb.FileMongoDbEntity
-
-@Repository
-interface FileMongoDbRepository : MongoRepository<FileMongoDbEntity, String> {
-    fun findByUserIdAndDeletedFalseOrderByCreatedDateDesc(userId: String): List<FileMongoDbEntity>
-    
-    fun findByIdAndUserIdAndDeletedFalse(id: String, userId: String): FileMongoDbEntity?
-}
