+++ src/main/kotlin/thatline/localup/user/entity/UserMongoDbEntity.kt
@@ -1,10 +1,10 @@
-package thatline.localup.user.entity.mongodb
+package thatline.localup.user.entity
 
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.core.index.Indexed
 import org.springframework.data.mongodb.core.mapping.Document
 import thatline.localup.common.constant.Role
-import thatline.localup.common.entity.mongodb.BaseMongoDbEntity
+import thatline.localup.common.entity.BaseMongoDbEntity
 import java.time.LocalDateTime
 
 @Document(collection = "user")
