+++ src/main/kotlin/thatline/localup/user/repository/BusinessMongoDbRepository.kt
@@ -1,6 +1,6 @@
-package thatline.localup.user.repository.mongodb
+package thatline.localup.user.repository
 
 import org.springframework.data.mongodb.repository.MongoRepository
-import thatline.localup.user.entity.mongodb.BusinessMongoDbEntity
+import thatline.localup.user.entity.BusinessMongoDbEntity
 
 interface BusinessMongoDbRepository : MongoRepository<BusinessMongoDbEntity, String>
