+++ src/main/kotlin/thatline/localup/user/repository/UserMongoDbRepository.kt
@@ -1,7 +1,7 @@
-package thatline.localup.user.repository.mongodb
+package thatline.localup.user.repository
 
 import org.springframework.data.mongodb.repository.MongoRepository
-import thatline.localup.user.entity.mongodb.UserMongoDbEntity
+import thatline.localup.user.entity.UserMongoDbEntity
 
 interface UserMongoDbRepository : MongoRepository<UserMongoDbEntity, String> {
     fun findByEmail(email: String): UserMongoDbEntity?
