+++ src/main/resources/application.yaml
@@ -73,8 +73,8 @@ etc-api:
       # 초단기실황조회
       get-ultra-srt-ncst:
         second-path: /getUltraSrtNcst
-        response-type: JSON
+        data-type: JSON
       # 예보버전조회
       get-fcst-version:
         second-path: /getFcstVersion
-        response-type: JSON
+        data-type: JSON
